Cloud Functions/Sandbox/requirements.txt
flask==1.1.2
google-cloud-storage==1.35.0
google-cloud-error-reporting==1.0.0
sqlalchemy
datetime
pg8000<=1.16.5
pandas==1.2.1
